projects/dynamicforms/src/components/currency/currency.component.spec.ts
import { Component, DebugElement, Input } from '@angular/core';
import { async, ComponentFixture, TestBed } from '@angular/core/testing';
import { FormControl, FormGroup, ReactiveFormsModule } from '@angular/forms';
import { By } from '@angular/platform-browser';
import { CurrencyService, I18nTestingModule, LanguageService } from '@spartacus/core';
import { DynamicFormsConfig } from '../../core/config/form-config';
import { FieldConfig } from '../../core/models/form-config.interface';
import { CurrencyComponent } from './currency.component';
import { of } from 'rxjs';
import { OccValueListService } from '../../occ/services/occ-value-list.service';

@Component({
  // tslint:disable
  selector: 'cx-error-notice',
  template: '',
})
class MockErrorNoticeComponent {
  @Input() warn;
  @Input() parentConfig;
}

class MockOccFormService {}
class MockLanguageService {
  getActive() {
    return of('en');
  }
}
class MockCurrencyService {
  getActive() {
    return of('EUR');
  }
}
const mockField: FieldConfig = {
  fieldType: 'currency',
  name: 'testCurrency',
  label: {
    en: 'Test Currency Label',
  },
};

const mockFormGroup = new FormGroup({
  testInput: new FormControl(),
});

const mockDynamicFormsConfig: DynamicFormsConfig = {
  dynamicForms: {},
};

describe('CurrencyComponent', () => {
  let component: CurrencyComponent;
  let fixture: ComponentFixture<CurrencyComponent>;
  let el: DebugElement;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [CurrencyComponent, MockErrorNoticeComponent],
      imports: [ReactiveFormsModule, I18nTestingModule],
      providers: [
        { provide: OccValueListService, useClass: MockOccFormService },
        { provide: LanguageService, useClass: MockLanguageService },
        { provide: CurrencyService, useClass: MockCurrencyService },
        {
          provide: DynamicFormsConfig,
          useValue: mockDynamicFormsConfig,
        },
      ],
    }).compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(CurrencyComponent);
    component = fixture.componentInstance;
    component.group = mockFormGroup;
    component.config = mockField;
    el = fixture.debugElement;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });
  it('should check components type', () => {
    expect(component.config).toBe(mockField);
    expect(component.config.fieldType).toEqual('currency');
  });

  it('should render input component', () => {
    const inputComponent = el.query(By.css('input[type="text"]')).nativeElement;
    expect(inputComponent).toBeTruthy();
  });
});
